[NP-52] fix: rebase
